backend/requirements.txt
Flask==3.0.3
flask-restx==1.3.0
mysql-connector-python==9.0.0
python-dotenv==1.0.1
